wave2c-4(housekeeping): clippy fixes + cargo fmt
* database.rs: drop redundant  cast on PreloadConfig::max_millis
  (already u64).
* cursor.rs: switch `mut config = DbiCfg::default(); config.x = ...`
  pattern to struct-update syntax to silence
  field_assignment_outside_initializer.
* cargo fmt: auto-applied formatting in noxu-xa, noxu-db cursor.

cargo clippy --workspace --all-targets is now clean of new warnings
introduced by Wave 2C-4.

+++ crates/noxu-db/src/database.rs
@@ -1142,7 +1142,7 @@ impl Database {
         // `elapsed_ms` to their config) but does not yet stop the
         // walk early.  Tracked for v2.0 alongside true LN warming.
         let elapsed_ms = start.elapsed().as_millis() as u64;
-        if max_millis > 0 && elapsed_ms > max_millis as u64 {
+        if max_millis > 0 && elapsed_ms > max_millis {
             log::warn!(
                 "Database::preload: walk took {elapsed_ms} ms, exceeding \
                  max_millis budget of {max_millis} ms (advisory until \
